refactored components after updating dashboardcontext file
