+++ client/src/pages/Dashboard.js
@@ -1,4 +1,4 @@
-import React, { useContext, useReducer } from "react";
+import React from "react";
 import { Grid, Paper } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
@@ -8,6 +8,7 @@ import Feed from "../components/Feed";
 import Ping from "../components/Ping";
 import ProfileBox from '../components/ProfileBox';
 import { useAuthContext } from '../utils/useAuthContext';
+import { useDashboardContext } from '../utils/useDashboardContext';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -23,96 +24,58 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const DashboardContext = React.createContext({
-  board: "",
-  details: "",
-  selectedUser: null,
-});
-
-const initialState = { board: "rawfeed" }
-
-function reducer(state, action) {
-  switch (action.type) {
-    case "rawfeed":
-      return {
-        ...state,
-        board: "rawfeed",
-        details: ""
-      }
-    case "ping":
-      return {
-        ...state,
-        board: "ping",
-        details: action.pingId
-      }
-    case "supportedpings":
-      return {
-        ...state,
-        board: "supportedpings"
-      }
-    default: 
-    return state;
-  }
-}
-
-export function useDashboardContext() {
-  return useContext(DashboardContext)
-}
-
 export default function Dashboard() {
-  const [state, dispatch] = useReducer(reducer, initialState);
   const classes = useStyles();
-  const context = useAuthContext();
+  const authContext = useAuthContext();
+  const [state] = useDashboardContext();
+
   return (
     <div className={classes.root}>
       <Nav />
       <div className={classes.grid}>
-        <DashboardContext.Provider value={[state, dispatch]}>
-          <Grid container spacing={2}>
-            <Grid
-              item
-              container
-              direction="column"
-              xs={4}
-              justify="space-between"
-            >
-              <Grid item>
-                <Paper
-                  style={{ backgroundColor: "#fcf8f2" }}
-                  className={classes.paper}
-                >
-                  {context.user ? (
-                    <ProfileBox />
-                  ) : (
-                      <SignUpOrIn />
-                    )}
-                </Paper>
-              </Grid>
-              <Grid item>
-                <Paper
-                  style={{ backgroundColor: "#fcf8f2" }}
-                  className={classes.paper}
-                >
-                  Map
-                </Paper>
-              </Grid>
+        <Grid container spacing={2}>
+          <Grid
+            item
+            container
+            direction="column"
+            xs={4}
+            justify="space-between"
+          >
+            <Grid item>
+              <Paper
+                style={{ backgroundColor: "#fcf8f2" }}
+                className={classes.paper}
+              >
+                {authContext.user ? (
+                  <ProfileBox />
+                ) : (
+                    <SignUpOrIn />
+                  )}
+              </Paper>
             </Grid>
-
-            <Grid item xs={8}>
-                <Paper
-                  style={{
-                    backgroundColor: "#fcf8f2",
-                    height: "80vh",
-                    overflow: "auto",
-                  }}
-                  className={classes.paper}
-                >
-                  {state.board === "ping" ? <Ping /> : <Feed />}
-                </Paper>
-                {/* {console.log(state)} */}
+            <Grid item>
+              <Paper
+                style={{ backgroundColor: "#fcf8f2" }}
+                className={classes.paper}
+              >
+                Map
+              </Paper>
             </Grid>
           </Grid>
-        </DashboardContext.Provider>
+
+          <Grid item xs={8}>
+            <Paper
+              style={{
+                backgroundColor: "#fcf8f2",
+                height: "80vh",
+                overflow: "auto",
+              }}
+              className={classes.paper}
+            >
+              {state.board === "ping" ? <Ping /> : <Feed />}
+            </Paper>
+          </Grid>
+        </Grid>
       </div>
     </div>
   );
